Save all delete and clear
