+++ src/style.css
@@ -44,9 +44,6 @@ body {
   display: flex;
   justify-content: space-between;
   cursor: move;
-  border: 1px solid gray;
-  margin: 0;
-  padding: 0;
 }
 
 .item.dragging {
